+++ modules/charts/.storybook/decorators.tsx
@@ -1,4 +1,4 @@
-import { ChartContainer } from '#components/helper/ChartContainer';
+import { ChartContainer } from '#components/ChartContainer';
 
 export const chartContainerDecorator = (Story) => (
 	<div style={{ height: '300px', width: '300px', display: 'flex' }}>
