+++ modules/charts/src/components/ChartsThemeProvider.tsx
@@ -1,5 +1,21 @@
 import { ComponentType, createContext, PropsWithChildren, useContext } from 'react';
 
+const defaultColors = [
+	// https://observablehq.com/@d3/color-schemes?collection=@d3/d3-scale-chromatic
+	'#a6cee3',
+	'#1f78b4',
+	'#b2df8a',
+	'#33a02c',
+	'#fb9a99',
+	'#e31a1c',
+	'#fdbf6f',
+	'#ff7f00',
+	'#cab2d6',
+	'#6a3d9a',
+	'#ffff99',
+	'#b15928',
+];
+
 interface ThemeProviderProps {
 	colors?: string[];
 	components?: {
@@ -23,7 +39,7 @@ export const ThemeContext = createContext<ThemeProviderProps | null>(null);
  * @returns JSX element that provides theme context to children
  */
 export const ChartsThemeProvider = ({ children, ...props }: PropsWithChildren<ThemeProviderProps>) => {
-	return <ThemeContext.Provider value={props}>{children}</ThemeContext.Provider>;
+	return <ThemeContext.Provider value={{ colors: defaultColors, ...props }}>{children}</ThemeContext.Provider>;
 };
 
 export const useThemeContext = () => {
