+++ modules/charts/src/components/charts/Bar/View.tsx
@@ -2,8 +2,8 @@ import { css } from '@emotion/react';
 import { ResponsiveBar } from '@nivo/bar';
 import { useMemo } from 'react';
 
+import { useThemeContext } from '#components/ChartsThemeProvider';
 import { useColorMap } from '#hooks/useColorMap';
-import { createColorMap } from '#theme/colors';
 import { BarChartProps } from './BarChart';
 import { arrangerToNivoBarChart } from './nivo/config';
 
@@ -11,12 +11,24 @@ interface BarChartViewProps {
 	data: any;
 	handlers: BarChartProps['handlers'];
 	theme: BarChartProps['theme'];
-	colorMap: any;
 }
 
-const colorMapResolver = ({ chartData }) => {
-	const keys = chartData.map(({ key }: { key: string }) => key); // specfic chart color map code
-	return createColorMap({ keys });
+/**
+ * Creates a chart color map for consistent colors across charts
+ * Uses closure-based state, not a React hook
+ **/
+const colorMapResolver = ({ chartData, colors }) => {
+	const colorMap = new Map<string, string>();
+
+	// used for "color wraparound" modulo
+	let colorIndex = 0;
+
+	chartData.forEach(({ key }) => {
+		const assignedColor = colors[colorIndex++ % colors.length];
+		colorMap.set(key, assignedColor);
+	});
+
+	return colorMap;
 };
 
 /**
@@ -25,9 +37,8 @@ const colorMapResolver = ({ chartData }) => {
  *
  * @param props - Bar chart view configuration
  * @param props.data - Transformed chart data in Nivo format
+ * @param props.handlers -
  * @param props.theme - Arranger theme configuration
- * @param props.colorMap - Color mapping for consistent chart colors
- * @param props.onClick - Optional click handler for chart interactions
  * @returns JSX element with responsive bar chart
  */
 export const BarChartView = ({ data, handlers, theme }: BarChartViewProps) => {
@@ -36,7 +47,9 @@ export const BarChartView = ({ data, handlers, theme }: BarChartViewProps) => {
 		: data;
 
 	// persistent color map
-	const { colorMap } = useColorMap({ chartData: sortedData, resolver: colorMapResolver });
+	const { colors } = useThemeContext();
+	console.log('bar colors', colors);
+	const { colorMap } = useColorMap({ chartData: sortedData, resolver: colorMapResolver, colors });
 
 	/**
 	 * TODO: improve "chart view" config/interface
