+++ modules/charts/src/components/charts/Bar/nivo/config.tsx
