--- modules/charts/src/constants.ts
@@ -1,1 +1,0 @@
-export const ARRANGER_MISSING_DATA_KEY = '__missing__';
