+++ modules/charts/src/main.tsx
@@ -1,6 +1,6 @@
-import { BarChart } from '#components/charts/BarChart/BarChart';
+import { BarChart } from '#components/charts/Bar/BarChart';
 import { SunburstChart } from '#components/charts/Sunburst/SunburstChart';
-import { ChartsThemeProvider } from '#components/theme/Provider';
+import { ChartsThemeProvider } from '#components/ChartsThemeProvider';
 import { HeadlessChart } from './components/Headless';
 
 import { ChartsProvider } from './components/Provider/Provider';
