--- modules/charts/src/stories/Charts.mdx
@@ -1,3 +1,0 @@
-## Charts
-
-high level documentation mdx that doesn't fit in with colocation of component
