--- modules/charts/src/theme/theme.mdx
@@ -1,1 +1,0 @@
-### Theming doc
