--- modules/charts/src/utils/mappings.ts
@@ -1,4 +1,0 @@
-// GQL field name to Arranger extended mapping JSON field name
-export const toJSONFieldName = (fieldName: string) => {
-	return fieldName.replaceAll('__', '.');
-};
