+++ modules/charts/src/utils/result.ts
